style: improve responsive style of header
