+++ pages/[[...hubPaths]].tsx
@@ -8,11 +8,11 @@ import { getHubWithPath } from "../lib/requests/hub/getHub";
 import { getHubs } from "../lib/requests/hub/getHubs";
 import { useFetchByHubId } from "../lib/useQueries";
 import { Footer } from "../ui/components/Footer";
-import { HeaderBar } from "../ui/components/Header";
 import AddHubModal from "../ui/components/HubModals/CreateHubModal/CreateHubModal";
 import Hero from "../ui/sections/Hero";
 import HubMenu from "../ui/sections/HubMenu";
 import LinkSection from "../ui/sections/LinkSection";
+import HeaderBar from "../ui/components/Header";
 
 export default function Home() {
   const router = useRouter();
