+++ ui/components/Header/index.ts
@@ -1,0 +1,3 @@
+import HeaderBar from "./Header";
+
+export default HeaderBar;
